--- Sources/ForageSDK/Foundation/Protocol/ForageDelegate.swift
@@ -1,12 +1,0 @@
-//
-//  File.swift
-//  
-//
-//  Created by Danny Leiser on 7/27/23.
-//
-
-public protocol ForageDelegate: AnyObject {
-    func focusDidChange(_ inputField: ForageUI)
-    func textFieldDidChange(_ inputField: ForageUI)
-    func blurDidChange(_ inputField: ForageUI)
-}
